fix(blueprint): remove unsupported variable type validation
With the introduction of custom types in the blueprint spec,
validation at the schema level for unsupported variable types is no
longer feasible as we need to allow for dynamic variable types
that can be validated after parsing the schema.
The main reason for this is that generally it is good practise to keep
encoding/decoding as clean as possible without requiring any outside
state or context and to support custom types, we need to have access
to resource providers which is hard to implement cleanly during
initial deserialisation of a blueprint as per the schema.