feat(blueprint): add complete feature set for validating core variables
Add the complete feature set for validating core (scalar) variables.
This includes the validation of integer, strings, floats and booleans.
This work includes validation that conforms to the blueprint specification
in the way it deals with validating allowed value lists (enums) and
default values along with ensuring user-provided runtime values
for variables are of the expected form.

This work also includes adding an Equal() method to scalar values
as a utility to make comparing scalar value wrappers comparable.

GitHubIssue: #13