feat(blueprint): add protobuf representation of an expanded blueprint
Adds a protobuf representation of an expanded blueprint
with generated code, this was added with the primary purpose
of performance optimisations to avoid having to parse every
substitution occurence in a blueprint for blueprints that
have not changed using an implementation of the
cache.BlueprintCache interface.
Using protobufs for serialisation of expanded blueprints
avoids complexity added to the existing JSON and YAML
serialisation functionality that serialises/deserialises
substitutions to and from strings, protobufs is also
a more efficient format for storage and transmission
to distributed caches for scalable implementations of
applications using the blueprint framework.