wip(blueprint-ls): add early experiments for building a language server
Adds early experimental code testing out using the glsp package
to build out a language server for the blueprint language that will
use the blueprint framework validation functionality and map the
results to LSP diagnostics to provide in-editor validation for
developers as per the blueprint specification.
This will only work well for simple rules in the specification that
are focused on syntax and semantics that can be caught at "initialisation"
time without requiring interaction with actual blueprint deployments
to understand the validity of the blueprint.