feat(blueprint-ls): add base json structures for lsp 3.17.0
Adds the base json structures as per the 3.17.0 LSP spec.
This also includes convenience functions for converting
character offsets from a negotiated position
encoding kind to byte offsets to be able to correctly
access the index of a string in Go as Go strings are
backed by byte arrays where slicing or accessing as
a specific location in a string with an index will be
a position of a byte in an underlying byte array.