feat(blueprint): add behaviour to track lines and columns when parsing yaml
The yaml.v3 parsing library allows us to track lines and columns of nodes
when implementing custom unmarshal behaviour for source yaml files.
The json parser built-in to go does not support this level of tracking
to keep it efficient and simple.
This limits what we can do when it comes to providing a great user experience
around editing blueprints with tools such as the language server when it comes
to json.
For yaml, however, we can track and expose line and column information at
the schema and specification level so when parsing yaml files, detailed
diagnostics including the locations in the source file can be provided
to the user.
This commit includes adding *meta.Source to components in the schema
and the first implementation with line information populated for
transforms and added to errors for the transform component.