feat(lib-rt-core): add websocket connection registry implementation
Add implementation of a WebSocket connection registry
that holds connections on the current instance and allows
broadcasting and listening to messages from other nodes in
a cluster.
The interface for a WebSocket connection registry to other
nodes in a cluster is a tokio mspc channel that is expected
to be connected to an external pub/sub network protocol
such as Redis pub/sub.

This also includes the wip implementation of the local runtime
api for inter-process communication between the runtime
and compiled exectuables that contain handler code.