refactor(blueprint): remove unnecessary /pkg/ directory
pkg is an indrection that is no longer seen as a best practise and most
go library developers do not use this convention.
The blueprint framework is purely a library with no executables so it does
not make sense to organise to distinguish between executables (cmd), internal
and public packages (pkg).