refactor(blueprint): change interfaces for provider and transformer
As the build engine provides a gRPC-based plugin system on top of the
Provider and SpecTransformer interfaces along with the child interfaces
for resources, data sources, links, custom variable types
and abstract resources, the initial interfaces were not sufficient.
To adapt from gRPC methods that require a context and can return an error
to simple method signature would require suppressing of important errors
that could lead to buggy plugins that do not report important errors.
For this reason, every method on these interfaces takes a context
and returns an error.
The blueprint framework is not tied to the implementation details of
the build engine plugin system but does require every operation that can be
carried out on these interfaces to take a context and return an error.

The re-design of these interfaces also ensures that for the Resource,
DataSource, CustomVariableType and Link interfaces, that every method
only takes the context and one other argument. This makes it easier to create
convenience wrappers in the build engine plugin SDK while using generics
to maintain type safety without additional runtime type assertions.