feat(blueprint): add support for values as per the latest blueprint spec updates
This adds support for value definitions and references in blueprints.
Values primarily exist to store computed values that need to be shared
between resources or exported from a blueprint, these can be seen as
local variables within a blueprint that can store the result of
a function applied to a value within a "${..}" substitution.