refactor(blueprint): expose schema from container and decouple loader from instance
Ensures the parsed blueprint schema is returned from loader
validate functions by exposing a BlueprintSpec method
on the container.
This work also includes a change that decouples the loader
from a specific blueprint instance; an instance ID was
required when constructing a loader to be passed into
the link function implementation of the core provider.
Instead of having to create a unique loader per blueprint,
the link function takes a callable to dynamically retrieve
the blueprint ID from the current context meaning a single loader
can be created for handling multiple blueprints.