refactor(blueprint): move element type checks to validation stage instead of parse stage
In order to enable features like completion suggestions for
value and data source field types, validation needed to be moved
as the blueprint needs to be parsed in order to provide suggestions
in the language server.
This also helps make the implementation more consistent, where all
semantic checks can be found in the validation sub-package
and the schema sub-package is purely about parsing a blueprint as per
the type schema along with parsing substitutions.