refactor(blueprint-ls): turn app into composition of language services
Refactors to a set of language servers that have access to
dependencies through the service instance struct instead of passing
around all the dependencies across many layers of functions.
This also helps define boundaries and will make for more maintainable
and testable code.

This also includes an implementation of the initial set of features
for intellisense completion suggestions.