refactor(blueprint): create ref chain collector for each load call
A ref chain collector holds state that is unique to each load call
in a blueprint loader. Initially, a single ref chain collector
was configured with a loader which worked fine for cases where
a loader is created per blueprint like in the language server.
In most use cases, the loader will be created once in an application
such as the Celerity build engine, this will then be used to load
multiple blueprints for the lifetime of the running application.
This change makes sure that a unique ref chain collector is
instantiated for each call to the load method.