fix(blueprint): add fix for intermittent bug with building link chains
A test was intermittently failing where a blueprint
with a hard cyclic link should always produce a
circular link error.
This behaviour turned out to be due to the link
selection logic allowing a resource to reference itself
in the "LinksTo" field of a chain link,
this lead to behaviour where the full chain link would
not be fully formed leading to a false positive on blueprints
that that should fail validation due hard cyclic links.