feat(blueprint): add functionality to group ordered resources to resolve concurrently
Adds logic to group an ordered chain of link nodes
for resources into groups of unrelated resources
which can be resolved in parallel.
Resolving in this case means deployment or
change staging.
This also includes some refactoring to rename link
and ref chain nodes, the previous naming could get confusing
for chain links, especially when dealing with standalone
resources that are not connected to others.