feat(blueprint): add depends on field to resource schema
Adds the "dependsOn" field to resources in the blueprint schema
to allow for explicit dependencies as per the latest iteration of the
blueprint spec.
This also refactors the transform values schema component to reuse
marshal/unmarshal logic for a list of strings that holds source location
information for source YAML documents.