wip(blueprint): work on change staging and clean up references to revisions
Adds work-in-progress pieces for change staging and
substitution resolution.
This also involves some refactoring work to remove the concept
of revisions for blueprint instances.
Managing revisions for state that shadows resources such
as cloud provider IaC resources does not make much sense
as unlike a piece of content or a document, switching between
revisions for an IaC blueprint would require reflecting
the switch by actioning the changes in the remote resource
providers, it is a much simpler approach to model simple deployments
with blueprint instance state reflecting the current snapshot
of the deployed blueprint and have the blueprint source documents
versioned and fed through the change staging and deployment
flow of the blueprint framework.