refactor(blueprint): update state container interface
Update state container interface to allow for more granular
interactions with the persisted state of a blueprint instance,
the idea is for other parts of the deployment "engine"
to use the state container interface as much as possible to reduce
the risk of bugs manipulating state structs behind pointers across
the board. This will also help with performance for state container
implementations that store sub-components of a blueprint instance
as their own tables and the InstanceState is a materialised view
exposed to components of the blueprint framework.