refactor(blueprint): combine spec and state for resource
Combines what was previously separate state and spec
definitions for a resource into a single attribute of a resource.
This simplifies change staging to compare incoming specs
for new deployments with the current spec stored as a part of
the current state of a blueprint instance.
This work adds new struct fields to the resource
definitions schema to allow for the definition of
computed fields in a spec, these are not defined by users
but instead derived from a resource provider after
a resource has been deployed.
An IDField property has also been added to the spec definition
to allow provider developers to specify which field should be resolved
to when a resource is referenced without accessing any of its properties.