test(blueprint-resolvers): add fix to ensure test container is removed between test runs
In the CI environment, we run the test suite twice,
once to collect code coverage and a second time to produce a JSON report
compatible with sonar.
Before, the second test run was failing as a part of the azure resolver
suite set up code, it provisions the blob container for the test.
This fix ensures the blob container is removed on tear down of
the test suite.