wip(blueprint): add refactoring validation and change staging work in progress
Adds refactoring of validation to support collecting the
property within an element primarily to be able to detect
which references are made from a resource `each` property
to check if the substitution in an `each` property
depends on resources or child blueprints.
This also adds the work in progress for change staging.