feat(blueprint): add functions to convert between mapping nodes and go types
Adds functions to convert between mapping nodes and Go values
when calling substitution functions as a part of substitution
resolution.
Functions in the blueprint framework are designed to interact with
Go types but the substitution engine uses the meta representation
of a value that is used in parsing substitutions.
This uses reflection for the conversion of Go values into
mapping nodes which will be quite slow, performance will only
be a concern if an instance of a host environment for the blueprint framework
is processing blueprints with lots of function calls with complex structures,
this assertion may prove to be incorrect in the future where an alternate
solution could be explored but for the initial implementation, reflection
seemed an appropriate choice.