refactor(deploy-engine): rename build engine to deploy engine
Rename build engine to deploy engine to reflect latest changes
to a better job and creating clearer separation of responsibilities
for build and deploy for a Celerity application.
The purpose of the deploy engine is to provide a layer on top
of the blueprint framework, packaging up core provider and persistence
layer implmenetations to manage the deployment lifecycle for the
infrastructure required for a Celerity application.
The deploy engine can also be used purely for the purpose of deploying
infrastruture-as-code defined in a blueprint.
Build steps for Celerity applications that can be written in multiple languages
that may involve compiling code and building docker images should be taken
care of by the CLI.