feat(blueprint): add validation for link function args
Adds validation for the special case where a resource reference
is provided as a string literal to the link function.
This logic ensures the reference is collected to be used for cycle
detection and in ordering elements for deployment at a later stage.
This work also includes some refactoring to use unified functions
to generate IDs used when collecting references to elements in
a blueprint.