test(blueprint): expand sub resolver test case for resources
Expands substitution resolver test case for resources
to include values that can only be resolved at deploy
time due to a dependency on an export from a child blueprint.
This also includes a fix to error handling to ensure
duplicate property paths do not appear in the final
resolve on deploy list.