fix(blueprint): allow any types for substitutions
Currently, validation is too strict about the resolved
types of substitutions.
The any type is used for all values decoded from a JSON string
where the actual type a value will not be known,
instead of preventing users from populating fields in a blueprint
with values decoded from JSON, there should be a more lenient approach
that allows for these values and assumes the blueprint developer
knows what they are doing and are certain that the values extracted
from a decoded JSON string are of the expected type.
Invalid types will be caught at runtime in this case, so there is some protection,
just not in the validation in a language server or in the CLI command that
the developer can run rapidly to get fast feedback on the validity of
a blueprint.