feat(blueprint): add implementation of resource change stager
Adds implementation of resource change stager that collects
the differences that will be applied by a new spec to the current
state of a resource in a deployed blueprint instance,
this also works for describing a "changes" for a resource that
will be deployed for the first time.
The resource change stager includes some protection for recursion
attack vectors by limiting traversal to max depth for mapping node
traversal used in validation and in the subengine when resolving
substitutions. This means there is a hard limit on the depth of
structures in a blueprint used in a resource spec or the custom metadata
section of a resource.