feat(blueprint): add helper function to extract values from mapping nodes
Adds a helper function to extract values by a simple path
similar to json path with limited syntax to allow extracting
values from a mapping node.
This was primarily created as a convenience function for link
implementations that access nested fields from the link data
modelled by a mapping node that differs from value extraction
used in substitution resolution as it is not tied to a resolve
context or resource specification and can easily be used in
other contexts.