refactor(blueprint): rename mapping node literals to scalars
The mapping node variant for scalar values was previously named Literal
which can be confusing as it can hold both literal and variables values
that happen to be strings, integers, booleans or floating point numbers.
This commit changes the name of the variant field to Scalar to better describe
the value that will be held in the struct field.

This work also includes some work-in-progress for a collection of helper functions
for creating link implementations along with an implementation of the stage changes
method for the test lambda function to dynamodb table link implementation.