fix(blueprint): add fixes for staging changes for brand new blueprint instances
Adds fixes for staging changes for a new blueprint instance,
including a test case.
This also includes refactoring duplicates of the same blueprint params
implementation across test utils, in the deploy engine and
language server to be a single implementation in the public API
of the core module that uses regular Go maps to store variables
under the hood.
This work also includes some logic to prevent blueprint include
cycles and set a max depth on a blueprint include tree, this work
is pending tests and possible fixes.