feat(blueprint): add behaviour to stage changes for destroying a blueprint instance
Add functionality to allow for "planning"
when destroying a blueprint instance.
This creates consistency with the inputs
destroy and deploy have to work with and for the workflow
to be carried out by users.
This includes alterations to the interfaces to allow an
intention to stage changes for destruction to be provided
and for the destroy method in a blueprint container
to take a set of removal changes to apply.