feat(blueprint): add method to check if resource has stabilised
Adds a method to the resource interface to check if a
resource has stabilised.
This is important for maximising efficiency in the deployment
process.
The Deploy method will complete as soon as the resource configuration
has been created and updated and the blueprint container will
periodically call HasStabilised if a resource must be stabilised
before a dependent can be deployed.