fix(blueprint): add support for a blueprint with includes but no resources
Adds support for blueprints that can contain includes
of other blueprints without any resources defined.
This allows for "wire-up" blueprints that are used to
wire together multiple blueprints for a complex stack
as a part of a modular design.
This commit also updates the version date to be the
anticipated date of launching v0.1.0 of Celerity
as an infrastructure-as-code tool.