feat(blueprint): add id generator interface and implementations
Adds an ID generator interface to allow the blueprint container to generate IDs
during the deployment process.
This also includes some core implementations including UUID, NanoID and
an empty ID generator.
The empty ID generator is useful for situations when the state container should
be responsible for assigning IDs, with an example being a state container
backed by a relational database.