feat(blueprint): add support for comparing against complex computed field paths
The link helpers provides convenience functionality for provider
developers to extract data from resource specs when staging
changes for a link.
Previously, the implementation only supported handling simple
computed fields as a part of collecting changes that are
dependent on resource fields.
This change ensures that more complex computed fields nested
in arrays and maps can be compared utilising the placeholders
for array items and map key-value pairs in the computed field
paths that are derived from a resource spec definition.