refactor(blueprint): extract deployment and change staging state to services
Instead of passing around mutable state and requiring
each function or part of the blueprint container to obtain
a lock on the state each time, it makes for a more maintainable
design in the long run to provide an abstracted interface
to interact with ephemeral state where the implementation
behind the interface is now responsible for synchronisation
across concurrent goroutines accessing the same state.
This also allows for concentrating state manipulation logic
under a more specific implementation rather than being spread
across utility functions and blueprint container methods.

This work also includes more work in progress for the
deployment implementation.