refactor(blueprint): extract resource change staging from blueprint container into its own service
This extracts the resource change staging logic that prepares for stage changing
and writes results to the change staging channels.
This includes renaming what was previous named the ResourceChangeStager to
ResourceChangeGenerator to distinguish between staging changes as a part
of a larger operation and generating a change set.
The ResourceChangeGenerator will also be used in the drift detection logic
which does not provide the same interface for reporting results as the
resource change stager.