fix(blueprint): add corrections to ordering deployment nodes
Adds corrections to behaviour to order deployment nodes
to using bubble sort to make sure every node is compared
with every other node to account for the fact that
dependencies are not always transitive and with an efficient
sort algorithm you can miss important comparisons that will
produce incorrect results in terms of the ordering of nodes
being deployed.
This also adds some refactoring to fix how references are
compared highlighted by an edge case revealed by the
higher level integration tests for the container deploy
method.