refactor(blueprint): move ref chains to separate package
Moves ref chains to a separate package
as they are now used heavily in the
deployment process in addition to
the validation stage.
This also includes changes to ensure that
all fields in a resource spec that are not
in the schema defined by the resource plugin
are caught and produce errors.