feat(blueprint): add drift check functionality for resources
Adds functionality that checks drift of resources in a blueprint
by utilising the resource plugin GetExternalState method
to get the live state of the resource in the upstream provider.
This is then compared against the state currently persisted
in the state container by reusing the change generator
used for change staging from which a difference is derived
and persisted.