refactor(blueprint): simplify redundant entity type prefixed field names in state
Simplifies redundant field names in resource,
instance and link state.
The exception to this change are the id fields
due to the ID name being used for the method
that fulfils the Element interface.
For this reason ResourceID, InstanceID and LinkID
will remain in the Go structs but will be serialised
to "id" when marshalling to JSON.