fix(blueprint): add missing fields to get external state input
In most cases, provider implementations should not need to interact with
the state container directly to get extra information.
This change makes sure that the currently known resource spec and metadata
is passed to the `GetExternalState` method for a resource as the third-party
ID of the resource will be in the spec data and annotations can be used to
provide extra information that decides the data that is retrieved.
Provider plugins could attach tags/metadata to the resource in the upstream
provider containing blueprint framework IDs and metadata but as this is not
always possible and takes control of the data away from the blueprint framework,
it is best if this information is passed to the provider methods.