feat(blueprint): add public protobuf conversion methods and improve provider errors
Makes some previously private protobuf conversion methods public
so that gRPC services (such as the deploy engine plugin system)
can reuse existing conversion behaviour.
This also adds a new provider.BadInput error type that can be used
to help distinguish between input and unexpected errors when
providing feedback to the user.
Bad input errors are not handled in any special way in the blueprint
container that manages deployments, it must be wrapped in specialised
errors to be handled correctly for deployment actions.
This also alters the public interface of existing provider errors
to make the error check functions more idiomatic where you write
to a target error and check the error type in the same action.