feat(blueprint): simplify resource registry interface for waiting for stability
Adds a convenience layer to the resource registry interface
and core implementation to allow callers to wait for a resource
to stabilise using a single boolean option instead of having
to set up a polling loop and call `HasStabilised` on an interval
to do so.
This is primarily useful for layers built on top of the blueprint
framework that allow the deployment of resources from outside
of a blueprint container (e.g. link plugins for deploying
intermediary resources).