feat(blueprint): add examples method for resources
Adds example method for resources to make plugin development
more documentation focused. This allows the core documentation
that would be generated in tooling for a plugin to be tightly
coupled with the implementation which means maintenance of
documentation for a plugin is the same as the implementation.
This also includes a new field in the resource definitions schema
for specifying examples for elements in a resource spec
definition.