feat(blueprint): enhance plugin interfaces with methods to provide rich documentation
Enhances plugin interfaces to include more documentation
so that rich documentation can be generated and tools
such as language servers can be further enhanced.
Having documentation as a part of the plugin interfaces
means that documentation and implementation should be maintained
together in the same place and makes for a better overall experience
for users of the blueprint framework and applications built on top of it.