refactor: reorganise deploy engine and plugin framework
Refactors the deploy engine to be a single project
and extracts the plugin framework into its own project.
The previous rationale was to create a library interface
for the deploy engine so it could be embedded directly
into the CLI. However, the deploy engine is essentially a
wiring up of a number of complex components which you
would usually find in an executable application.
Parts of the deploy engine can still be used by other projects
as there will be a lot of reusable components that will sit
outside of the internal folder to allow for bespoke distributions
or custom wire-ups in to embed the engine into
alternative versions of the Celerity CLI.
The previously separate apps/api project was going to be a thin layer
providing an API server interface for the deploy engine,
this has now beem unified into the single deploy engine sub-project.
It also made sense to decouple the plugin framework from the deploy
engine project as even at this early stage it has started to become
a sizeable project, separating concerns allows the plugin framework
to be managed and versioned separately from the deploy engine
application and creates a clearer boundary when the plugin framework
is used in other projects (such as the language server).