feat(plugin-framework): add complete implementation of top-level provider plugin methods
Adds the complete implementations of the top-level methods
of a provider plugin.
This also adds updates to the gRPC interface for plugins
to include the host ID in requests and the functionality
in provider plugins to check the host ID for a request
matches the one that the plugin was initially registered with.