feat(blueprint): add with call method to function call context
This adds a WithCall method to the function call context
to allow for programmatic function calls between
provider functions.
This primarily facilitates for uses like that of the plugin framework
that will make function-to-function calls over an inter-process
call, where a function would send a copy of the call stack
with the additional function "over-the-wire" to the other plugin
function without modifying the existing stack in the parent context.