fix(blueprint): remove with call method and make call stack thread-safe
Removes the WithCall method that was added to the
plugin function call context as it was added due to
a forgotten understanding of how the plugin service
works in the plugin framework.
As on the server-side (plugin host, e.g. deploy engine)
it will call the default function registry which handles
pushing new function calls to the stack and popping
them off when the call is fininshed or has timed out.
Due to this, there is no need to send a copy of the call stack
with the extra function between the plugin and the
plugin service.
This commit also includes a fix to make it so the
function call stack for plugins can be accessed
across multiple goroutines.