feat(plugin-framework): add support for keying transformer plugins by namespace
Adds option to the plugin launcher to key
transformers by plugin name (namespace)
instead of the transform string used in blueprints.
This allows tools like the plugin JSON doc
generator to identify transformer plugins
in a transform.SpecTransformer map based on the plugin
ID when all the information it has is the plugin ID
before launching.
Keying by transform string is kept as an option to allow
for the blueprint loading use case where the only information
a blueprint loader has for a transform plugin is the
transform string used in the transform section of a blueprint.