feat(blueprint): add support for a blueprint instance name
Adds a new method to the state container to look up
an instance ID given a name.
This allows users to assign and use their own unique
names for deploying, destroying and staging changes
for blueprint instances.
This commit includes changes across the `Deploy`,
`Destroy` and `StageChanges` methods for a blueprint
container to accept a blueprint instance name
and logic to resolve an instance ID to reduce the surface
area of changes that need to be made to facilitate
accepting a name.