feat(blueprint-resolvers): make change to use default credentials for azure
Makes a change to ensure that for non-test environments
that the Azure credentials are sourced from the environment
or credential files on the host machine using the SDKs
default credentials.
This also ensures that a client is created on every resolve
request to allow for sourcing client-specific configuration
or credentials dynamically for include metadata or from
blueprint params context variables.