feat(plugin-framework): add support for a different timeout for resource stabilisation
Adds support to use a different timeout for plugin to plugin calls
where a plugin requests to deploy a resource and wait until the
resource has stabilised.
The default plugin to plugin call timeout is supposed to be quite
short to prevent infinite waits and infinite recursion due to
erroneous plugin implementations, however, resource stabilistion
can take a long time so an alternative timeout is needed where issues
with erroneous recursive calls is more likely to come from plugin
functions.