fix(blueprint): add fix to avoid panicking when state container is nil
Add fix to avoid panicking when the state container is nil
and the "core" provider has not been provided by the caller.
This adds a test case for the bug and a solution that
introduces a no-op links container that can be used to
instantiate the core provider when in a context that
should not require a state container.
When validating blueprints in an application such as the
blueprint language server, there is no need for a state
container as the language server does not orchestrate
deployments. The core provider, however, requires
a links container for the "link" function
so to allow the creation of the core provider in validation
only contexts that use methods to get schemas and docs
for functions, a no-op links container has been introduced
so that any calls made to the core link function in a context
without a state container will not cause the host
application to panic.