fix(blueprint-state): add memfile fix for saving events in-memory
Adds fix to the memfile state container
implementation to make sure that the event is written
to the in-memory partition slice.
This fixes a bug where no events were being sent to the stream
after saving an event until the host application was reloaded
to re-build the in-memory state from persisted files.