fix(blueprint-state): make sure end field is copied for events
When copying events for persistence in the `memfile`
state container implementation, the `End` value was not being
copied meaning that no end of stream marker was persisted to file.
This would lead to streams not closing until the client requests
to end after waiting for too long without receiving any events.