fix(plugin-framework): correct protocol compatibility checks
Correct protocol compatibility checks so that
the host protocol version and plugin protocol
versions do not require an exact match to be
compatible.
Any minor plugin protocol version that is
less than or equal to the host will be compatible.
For example, a plugin that supports protocol "1.2"
will be compatible with a host using protocol version
"1.4". The behaviour remains the same for major version
differences, the host is not compatible with plugins
that do not support the major version of the host.
Plugins that only support a minor version that
comes after the host minor version are not compatible,
for example a plugin that supports protocol "1.12"
will not be compatible with a host using protocol version
"1.5".